crates/rust-daq/clients/python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel", "grpcio-tools>=1.50"]
build-backend = "setuptools.build_meta"

[project]
name = "rust-daq-client"
version = "0.1.0"
description = "Python client library for rust-daq headless daemon"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "rust-daq contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "grpcio>=1.50",
    "grpcio-tools>=1.50",
    "protobuf>=4.20",
    "anyio>=3.0",
    "numpy>=1.20",
]

[project.optional-dependencies]
scan = [
    "pandas>=1.3",
    "tqdm>=4.60",
]
jupyter = [
    "ipywidgets>=7.6",
    "notebook>=6.4",
    "matplotlib>=3.3",
    "plotly>=5.0",
    "tqdm>=4.60",
]
cocoindex = [
    "cocoindex[embeddings]>=0.1",
    "python-dotenv>=1.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-mock>=3.10",
    "black>=23.0",
    "mypy>=1.0",
    "ruff>=0.1",
    "pdoc>=14.0",
]
docs = [
    "pdoc>=14.0",
]
all = [
    "rust-daq-client[scan,jupyter,cocoindex,dev,docs]",
]

[project.urls]
Homepage = "https://github.com/yourusername/rust-daq"
Documentation = "https://github.com/yourusername/rust-daq/blob/main/clients/python/README.md"
Repository = "https://github.com/yourusername/rust-daq"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]
markers = [
    "integration: marks tests that require a running rust-daq daemon (deselect with '-m \"not integration\"')",
]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pdoc]
# Output directory for generated docs
output_directory = "../../docs/reference/python"
# Docstring format (Google style)
docformat = "google"
# Modules to document (exclude generated protobuf code)
modules = ["rust_daq"]
# Template directory (optional)
# template_directory = "docs/templates"
